validacao numero negativo/saldo como double

+++ BackEnd/BackEnd/Controllers/TransacaoController.cs
@@ -114,7 +114,7 @@ namespace BackEnd.Controllers
                         id = Convert.ToInt32(reader["id"]),
                         id_tipo_conta = Convert.ToInt32(reader["id_tipo_conta"]),
                         id_pessoa = pessoa,
-                        saldo = Convert.ToInt32(reader["saldo"])
+                        saldo = Convert.ToDouble(reader["saldo"])
 
                     };
